Refactor logout configuration and update dev.properties
This commit focuses on two significant updates. The logout configuration in the SecurityConfiguration file has been refactored to a lambda styled structure to enhance readability and consistency with the codebase.

In the dev.properties file, the sequence of bot.token and bot.username properties is reordered to improve alike properties' grouping. Also, the database URL has been updated from a hardcoded IP to localhost to enhance the application's adaptability in different environments.

These changes will boost code maintainability without affecting the program's functionality.